+++ FirebaseVertexAI/CHANGELOG.md
@@ -6,6 +6,8 @@
   Note: This feature is in Public Preview and relies on experimental models,
   which means that it is not subject to any SLA or deprecation policy and could
   change in backwards-incompatible ways.
+- [added] Added support for specifying the minimum and maximum number of items
+  (`minItems` / `maxItems`) to generate in an array `Schema`. (#14671)
 
 # 11.11.0
 - [added] Emits a warning when attempting to use an incompatible model with